pyproject.toml
[project]
name = "claudestreet"
version = "0.1.0"
description = "AWS-native serverless AgentSwarm for autonomous stock trading — inspired by OpenClaw"
requires-python = ">=3.12"
license = "MIT"
dependencies = [
    "anthropic>=0.40",
    "boto3>=1.34",
    "pydantic>=2.5",
    "pyyaml>=6.0",
    "numpy>=1.26",
    "pandas>=2.1",
    "ta>=0.11",
    "yfinance>=0.2",
    "feedparser>=6.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.23",
    "moto[dynamodb,events,s3,ssm,secretsmanager]>=5.0",
    "ruff>=0.1",
]
cdk = [
    "aws-cdk-lib>=2.120",
    "constructs>=10.0",
]
broker = ["alpaca-trade-api>=3.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/claudestreet"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
